SuperComputer and Human Logic
Supercomputer will generate an array with the values of 2^n - 1 up to the nth value which is higher than the number of marbles which is the upper limit.
Then from there, it will get the number less than the number of marbles and subtract that value from the pile to always leave the pile with a value of 2^n - 1 which will lead to victory for the super computer.